# src/main/resources/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "alojamiento quindio",
    "description": "API REST para la gestión completa",
    "contact": {
      "name": "Equipo de Desarrollo",
      "url": "https://tiendavirtual.com",
      "email": "dev@tiendavirtual.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Servidor de Desarrollo"
    },
    {
      "url": "https://api-test.tiendavirtual.com",
      "description": "Servidor de Pruebas"
    },
    {
      "url": "https://api.tiendavirtual.com",
      "description": "Servidor de Producción"
    }
  ],
...
  "paths": {
    "/api/v1/host/bookings/{id}/cancel": {
      "put": {
        "tags": [
          "Host Bookings"
        ],
        "summary": "Cancelar reserva por anfitrión",
        "description": "Solo para anfitrión propietario; según políticas, genera notificación.",
        "operationId": "cancelBookingByHost",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la reserva",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        ],
        "responses": {
          "200": {
            "description": "Reserva cancelada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDTO"
                }
              }
            }
          },
          "400": {
            "description": "Fuera de plazo",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDTO"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDTO"
                }
              }
            }
          },
          "403": {
            "description": "No propietario",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDTO"
                }
              }
            }
          },
          "404": {
            "description": "No encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/host/bookings/{id}/approve": {
      "put": {
        "tags": [
          "Host Bookings"
        ],
        "summary": "Aprobar reserva",
        "description": "Solo para anfitrión propietario; cambia estado a CONFIRMED, genera notificación.",
        "operationId": "approveBooking",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la reserva",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        ],
        "responses": {
          "200": {
            "description": "Reserva aprobada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDTO"
                }
              }
            }
          },
          "400": {
            "description": "Estado inválido",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDTO"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDTO"
                }
              }
            }
          },
          "403": {
            "description": "No propietario",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDTO"
                }
              }
            }
          },
          "404": {
            "description": "No encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/host/dashboard": {
      "get": {
        "tags": [
          "Host Dashboard"
        ],
        "summary": "Obtener resumen del dashboard del anfitrión",
        "description": "Incluye ingresos, calificaciones promedio, reservas pendientes y métricas de propiedades.",
        "operationId": "getHostDashboard",
        "responses": {
          "200": {
            "description": "Resumen obtenido",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummaryDTO"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummaryDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/host/bookings": {
      "get": {
        "tags": [
          "Host Bookings"
        ],
        "summary": "Listar reservas de los alojamientos del anfitrión",
        "description": "Historial con filtros. Paginado.",
        "operationId": "getHostBookings",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Estado de la reserva",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "PENDING"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Número de página",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "example": 0
          },
          {
            "name": "size",
            "in": "query",
            "description": "Tamaño de página",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            },
            "example": 10
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de reservas",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingDTO"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingDTO"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/dashboard": {
      "get": {
        "tags": [
          "Main Dashboard"
        ],
        "summary": "Obtener resumen del dashboard principal",
        "description": "Estadísticas globales del sistema: usuarios, reservas, ingresos totales, etc.",
        "operationId": "getMainDashboard",
        "responses": {
          "200": {
            "description": "Resumen obtenido",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalDashboardSummaryDTO"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalDashboardSummaryDTO"
                }
              }
            }
          },
          "403": {
            "description": "Rol no autorizado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalDashboardSummaryDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "BookingDTO": {

      },
...
      "DashboardSummaryDTO": {

      },
      "GlobalDashboardSummaryDTO": {

      }
    }
  }
}